FEAT: Lesson Planner Tool updates (#125)
* FIX: Error on Curriculum delete

When curriculum is assigned to a classroom and then deleted, page breaks because of invalid id assoc. Temp fix to gracefully show error instead of breaking

* FEAT: Column pinning in advanced table

* FEAT: Improved Curriculum list index

* FIX: Typescript fix for Advanced Table column def

* TEST: Curriculum delete tests updated

* FEAT: Delete check on curriculum delete

* FIX: Updated

+++ app/Http/Controllers/CurriculumController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\CurriculumPutRequest;
 use App\Models\Curriculum;
 use Illuminate\Http\Request;
 use App\Models\Student;
+use App\Models\Classroom;
+use Illuminate\Support\Facades\Log;
 use Inertia\Inertia;
 use stdClass;
 
@@ -180,6 +182,16 @@ class CurriculumController extends Controller
      */
     public function destroy(Curriculum $curriculum)
     {
+        $classroomsCollection = Classroom::where('curriculum_id', $curriculum->id)->get();
+        if ($classroomsCollection->isNotEmpty()) {
+            Log::warning(
+                "Curriculum ({$curriculum->name}) has already been assigned to school classrooms. Cannot be deleted as it will break relations.",
+                [
+                    "Classroom names" => $classroomsCollection->value('name'),
+                ]
+            );
+            return redirect()->route('curriculum.index')->with('warning', 'Curriculum assigned to classroom. Cannot be deleted');
+        }
         Curriculum::findOrFail($curriculum->id)->delete();
 
         return redirect()->route('curriculum.index')->with('success', 'Curriculum deleted successfully');
